Move deployable_entity fixture to root
