+++ client/verta/tests/conftest.py
@@ -20,6 +20,9 @@ import verta
 from verta import Client
 from verta._internal_utils import _utils, _pip_requirements_utils
 from verta.environment import Python
+from verta.tracking.entities._deployable_entity import _DeployableEntity
+from verta.tracking.entities import ExperimentRun
+from verta.registry.entities import RegisteredModelVersion
 
 import hypothesis
 import pytest
@@ -465,6 +468,26 @@ def registered_model(client, created_entities):
 @pytest.fixture(scope="class")
 def class_registered_model(class_client, class_created_entities):
     return registered_model_factory(class_client, class_created_entities)
+
+
+@pytest.fixture(params=utils.sorted_subclasses(_DeployableEntity))
+def deployable_entity(request, client, created_entities):
+    cls = request.param
+    if cls is ExperimentRun:
+        proj = client.create_project()
+        created_entities.append(proj)
+        entity = client.create_experiment_run()
+    elif cls is RegisteredModelVersion:
+        reg_model = client.create_registered_model()
+        created_entities.append(reg_model)
+        entity = reg_model.create_version()
+    else:
+        raise RuntimeError(
+            "_DeployableEntity appears to have a subclass {} that is not"
+            " accounted for in this fixture".format(cls)
+        )
+
+    return entity
 
 
 @pytest.fixture
